requirements.txt
requests
feedparser
google-generativeai
python-dotenv
beautifulsoup4
schedule
